modify browserlisting as a component
